integration: rename session.rs -> db_errors.rs
The remaining two tests in session.rs:
- test_db_errors(),
- test_rate_limit_exceeded_exception(),
regard DB errors, so the file is renamed accordingly.

This finishes dissolution of the old overloaded `session_test.rs` module.
Hooray!
