+++ scylla/tests/integration/session/mod.rs
@@ -1,11 +1,10 @@
 mod authenticate;
 mod caching_session;
+mod db_errors;
 mod history;
 mod new_session;
 mod pager;
 mod retries;
 mod self_identity;
-#[expect(clippy::module_inception)]
-mod session;
 mod tracing;
 mod use_keyspace;
